Warning if you try to leave with unsaved changes.
We need to impelment this ourselves, rather than using the standard
formchangechecker, because

a) our code has more detailed knowledge of what is going on, and can
handle unsaved changes coming and going. (formchangechecker can only
cope with a form going unchanged -> changed -> submitted. Not going back
to unchanged, then listening for future changes.

b) we want to be able to Behat test this.